refactor(game-engine): extract getLegalMoves to legal-moves.ts (S27.8.11)
Extract `getLegalMoves` (~244 lines) and its private helper variant of
`getAdjacentOpponents` (~25 lines) from the monolithic `actions.ts`
dispatcher to a dedicated `actions/legal-moves.ts` module.

The local variant of `getAdjacentOpponents` only filters `stunned`
players, contrary to the public version in `mechanics/movement.ts`
which also excludes hypnotized and `titchy` players. This behavior
is preserved verbatim by keeping the helper private to the new module.

Re-export from `actions.ts` to preserve the public API consumed by
`index.ts`, `ai/difficulty.ts`, `ai/evaluator.ts`, `utils/referee.ts`,
and several test suites. No cycle back to `actions.ts`: all helpers
come from `mechanics/*` or `core/*`.

Massive orphan import cleanup in `actions.ts`:
- Direct consequences of this extraction: `Player`, `TeamId`,
  `inBounds`, `canPlayerMove`, `canPlayerContinueMoving`,
  `shouldAutoEndTurn`, `canTeamBlitz`, `getPassRange`,
  `canAttemptPassForRange`, `isAdjacent`, `canThrowTeamMate`,
  `getThrowRange`, `canHypnoticGaze`, `canProjectileVomit`,
  `canStab`, `canChainsaw`, `canLeap as playerCanLeap`,
  `getLegalLeapDestinations`, `canInstablePerformAction`,
  `logInstablePrevention`.
- Pre-existing orphans cleaned up opportunistically:
  `applyPickupFailure`, `canDumpOff`, `getDumpOffReceivers`,
  `checkDauntless`, `getArmBarBonus`.

`actions.ts` shrinks 1165 -> 896 lines (-269 net lines, cumulative
-1810 vs origin 2706 = ~67% reduction toward the DoD target of <= 600).

9 new smoke tests in `legal-moves.test.ts` cover the exported API
surface and short-circuits (`gamePhase===ended`, missing `players`,
`isTurnover`, `pendingApothecary`, `pendingReroll`,
`pendingKickoffEvent`, `END_TURN` fallback, state immutability).

- 4910/4910 game-engine tests pass (+9 new smoke tests)
- 1787/1787 server tests pass (no regression)
- typecheck + lint + build clean
